moved new request modal code to separate file
