+++ services/request.js
@@ -39,7 +39,6 @@ export const getRequest = async (userID, id) => {
 };
 
 export const createRequest = async (request) => {
-  console.log('create request services');
   try {
     const token = await AsyncStorage.getItem('authToken');
     const { data } = await axios.post(`${URL}`, request, { headers: { Authorization: `Bearer ${token}` } });
